auth: Plugins can now handle responses directly.
Authentication plugins can now directly handle the HTTP response body,
headers, and code directly if wanted. To handle the responses within the
plugin start_response must be called and a list (the response body) must
be returned. If the plugin wants to allow the plugin system to handle
the rejection the plugin should continue to return True on success
and False on failed authentication.

Example:

   def authenticate(self, environ, start_response):
       start_response('200 OK', [('content-type', 'text/plain')])
       return [bytes('example', 'utf8')]