dispatcher: Bus injection now easier and enforced.
Before dispatch can be called _bus must be set. Previously there was a
check that attempted to create a new instance if _bus had not been set.
This change removed the Bus creation and, instead, makes it easier to
pass in a Bus connection and raises during Dispatch unless _bus is set.