handlers: Add decorators to handler functions.
The Dispatcher's dispatch() method is doing too much work, and is
assuming the entire Commissaire REST API is JSON-RPC based.  I think
this assumption is going to bite us in the long-term as it almost did
with the (now shelved) Tang passthrough API.

This adds a JSONRPC_Handler class to be used as a decorator for HTTP
request handler functions.  It takes over the pre and post-processing
work around the handler call that dispatch() was previously doing.

This also adds a BasicHandler class from which JSONRPC_Handler class
inherits.  For now, this mainly serves as a tag for non-JSON-RPC handler
functions, so all handler functions can be identified without guesswork.