auth: Add self-auth option
AuthenticationManager can now be handed a self-auth list. Each item in
this list denotes a local HTTP/S path that provides it's own
authN/authZ. The AuthenticationManager will pass the request directly
to the endpoint without using Commissaire's authentication stack.

The option can be set via the command line:

    $ commissaire-http [...] \
      --self-auth /some/path \
      --self-auth /anotherpath

or by adding a "self_auth" to the configuration file:

   [...]
   "self_auth": ["/some/path", "/anotherpath"],
   [...]