Require bcrypt instead of py-bcrypt.
Custodia brings in bcrypt, but unfortunately both bcrypt and py-bcrypt
install a "bcrypt" module and they are slightly incompatible, causing
HTTPBasicAuth to break if it happens to import the wrong bcrypt module.

Instead of trying to work around that, just switch back to the bcrypt
package.  Steve says he likes it better than py-bcrypt anyway.